feat(server): WIP - save local change
