+++ packages/server/config/envs.ts
@@ -1,0 +1,4 @@
+import { resolve } from 'node:path'
+import { config } from 'dotenv'
+
+config({ path: resolve('..', '..', '.env') })
