+++ packages/server/lib/utils.ts
@@ -1,0 +1,7 @@
+export function delay(milliseconds: number): Promise<void> {
+  return new Promise(resolve => {
+    setTimeout(() => {
+      resolve()
+    }, milliseconds)
+  })
+}
