+++ packages/server/package.json
@@ -10,7 +10,6 @@
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@types/morgan": "^1.9.9",
     "cors": "^2.8.5",
     "express": "^4.19.2",
     "groq-sdk": "^0.5.0",
@@ -18,6 +17,7 @@
   },
   "devDependencies": {
     "@types/cors": "^2.8.17",
-    "@types/express": "^4.17.21"
+    "@types/express": "^4.17.21",
+    "@types/morgan": "^1.9.9"
   }
 }
